Issue 148 cleanup create repair request form for event manager (#159)
* change the repair-request form

* slight change on formatting

* fix formatting

* make create-repair-request as a component, reuse in /repair-request/index.tsx

* cleaning comments

* fixed one dependency issue

* modified the create-repair-request form to take an eventId as a prop

* removing headers in the form component

* passing eventId if the dropdown isn't rendered as default value to create-repair-request form on src/components/Forms/create-repair-request.tsx

* undo the changes on .env.example
